backend/pyproject.toml
[project]
name = "retriever"
version = "0.1.0"
description = "RAG platform with FastAPI and authentication"
authors = [
    {name = "retriever.sh", email = "support@retriever.sh"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "psycopg[binary]",
    "pydantic",
    "pydantic['email']",
    "pydantic-settings",
    "python-jose[cryptography]",
    "python-multipart",
    "passlib[bcrypt]",
    "boto3",
    "aiofiles",
    "httpx",
    "anyio",
    "numpy",
    "huggingface-hub",
    "llama-cpp-python",
    "polar-sdk",
    "logfire",
    "logfire[sqlalchemy]",
    "logfire[fastapi]",
    "pytest>=8.4.2",
]
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
